Web Socket Fully works

+++ src/main/java/com/example/collections_backend/collectionItem/itemLikes/LikeRepository.java
@@ -12,6 +12,6 @@ public interface LikeRepository extends JpaRepository<LikeItem, Long> {
 
     boolean existsByCollectionItemAndUser(CollectionItem collectionItem, User user);
 
-    Long countAllByCollectionItem(CollectionItem collectionItem);
+    int countAllByCollectionItem(CollectionItem collectionItem);
 
 }
